rejoining the streak

+++ src/codeForces/WOWFactor.java
@@ -5,13 +5,13 @@ public class WOWFactor {
 		Scanner scan = new Scanner(System.in);
 		String s = scan.next();
 		
-		int WFrontcnt = 0 ;
-		int WBackcnt = 0;
+		long WFrontcnt = 0 ;
+		long WBackcnt = 0;
 		
 		for(int i = s.length() - 1 ; i > 0 ; i--) if(s.charAt(i) == 'v' && s.charAt(i - 1) == 'v')WBackcnt++;
 //		System.out.println(WBackcnt);
 		long ans = 0;
-		int vCnt = 0;
+		long vCnt = 0;
 		for(int i = 0 ; i < s.length() ; i++) {
 			if(s.charAt(i) == 'o') {
 				WFrontcnt += (vCnt > 1 ? vCnt - 1 : 0);
